+++ test/OpenTelemetry.Contrib.Shared.Tests/SqlConnectionDetailsTests.cs
@@ -1,8 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using Xunit;
-
 namespace OpenTelemetry.Instrumentation.Tests;
 
 public class SqlConnectionDetailsTests
@@ -20,6 +18,10 @@ public class SqlConnectionDetailsTests
     [InlineData("tcp:localhost", "localhost", null, null, null)]
     [InlineData("tcp:[::1]", null, "[::1]", null, null)]
     [InlineData("tcp : localhost", "localhost", null, null, null)]
+    [InlineData("tcp://some.domain.local:5432", "some.domain.local", null, null, 5432)]
+    [InlineData("tcp://some.domain.local", "some.domain.local", null, null, null)]
+    [InlineData("tcp://[::1]:5432", null, "[::1]", null, 5432)]
+    [InlineData("tcp://[::1]", null, "[::1]", null, null)]
     [InlineData("np : localhost", "localhost", null, null, null)]
     [InlineData("lpc:localhost", "localhost", null, null, null)]
     [InlineData("np:\\\\localhost\\pipe\\sql\\query", "localhost", null, null, null)]