+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/EventSourceTests.cs
@@ -3,7 +3,6 @@
 
 using OpenTelemetry.Instrumentation.AspNetCore.Implementation;
 using OpenTelemetry.Tests;
-using Xunit;
 
 namespace OpenTelemetry.Instrumentation.AspNetCore.Tests;