+++ src/OpenTelemetry.Instrumentation.Wcf/CHANGELOG.md
@@ -18,6 +18,9 @@
   `ExecutionContext` flow was suppressed.
   ([#4378](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4378))
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+  
 * Add support for version [1.41.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.41.0/docs/rpc/README.md)
   of the Semantic Conventions for RPC/gRPC when the `OTEL_SEMCONV_STABILITY_OPT_IN`
   environment variable is set to `rpc` or `rpc/dup`.